+++ dart/geobase/lib/src/vector/content/geometry_content.dart
@@ -4,6 +4,8 @@
 //
 // Docs: https://github.com/navibyte/geospatial
 
+import '/src/coordinates/base/box.dart';
+
 import 'simple_geometry_content.dart';
 
 /// A function to write geometry data to [output].
@@ -15,8 +17,8 @@ typedef WriteGeometries = void Function(GeometryContent output);
 /// [SimpleGeometryContent] and geometry collections. It's possible that in
 /// future versions other geometry types are added.
 ///
-/// Coordinate positions, position arrays and bounding boxes are represented as
-/// coordinate value arrays of `Iterable<double>`.
+/// Coordinate positions and position arrays are represented as coordinate value
+/// arrays of `Iterable<double>`.
 mixin GeometryContent implements SimpleGeometryContent {
   /// Writes a geometry collection from the content provided by [geometries].
   ///
@@ -26,15 +28,7 @@ mixin GeometryContent implements SimpleGeometryContent {
   /// Use an optional [name] to specify a name for a geometry (when applicable).
   ///
   /// An optional [bounds] can used set a minimum bounding box for a geometry
-  /// written. A writer implementation may use it or ignore it. Supported
-  /// coordinate value combinations by coordinate type:
-  ///
-  /// Type | Expected values
-  /// ---- | ---------------
-  /// xy   | minX, minY, maxX, maxY
-  /// xyz  | minX, minY, minZ, maxX, maxY, maxZ
-  /// xym  | minX, minY, minM, maxX, maxY, maxM
-  /// xyzm | minX, minY, minZ, minM, maxX, maxY, maxZ, maxM
+  /// written. A writer implementation may use it or ignore it.
   ///
   /// An example to write a geometry collection with two child geometries:
   /// ```dart
@@ -60,6 +54,6 @@ mixin GeometryContent implements SimpleGeometryContent {
     WriteGeometries geometries, {
     int? count,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   });
 }