+++ dart/geobase/lib/src/vector/formats/wkb/wkb_encoder.dart
@@ -52,7 +52,7 @@ class _WkbGeometryEncoder
     Iterable<double> chain, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     // write a line string geometry
     _writeGeometryHeader(Geom.lineString, type);
@@ -64,7 +64,7 @@ class _WkbGeometryEncoder
     Iterable<Iterable<double>> rings, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     // write a polygon geometry
     _writeGeometryHeader(Geom.polygon, type);
@@ -83,7 +83,7 @@ class _WkbGeometryEncoder
     Iterable<Iterable<double>> points, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     // write a multi point geometry
     _writeGeometryHeader(Geom.multiPoint, type);
@@ -103,7 +103,7 @@ class _WkbGeometryEncoder
     Iterable<Iterable<double>> lineStrings, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     // write a multi line geometry
     _writeGeometryHeader(Geom.multiLineString, type);
@@ -123,7 +123,7 @@ class _WkbGeometryEncoder
     Iterable<Iterable<Iterable<double>>> polygons, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     // write a multi polygon geometry
     _writeGeometryHeader(Geom.multiPolygon, type);
@@ -150,7 +150,7 @@ class _WkbGeometryEncoder
     WriteGeometries geometries, {
     int? count,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     final int numGeom;
     final Coords coordType;
@@ -289,7 +289,7 @@ class _GeometryCollector with GeometryContent {
     Iterable<double> chain, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     hasZ |= type.is3D;
     hasM |= type.isMeasured;
@@ -301,7 +301,7 @@ class _GeometryCollector with GeometryContent {
     Iterable<Iterable<double>> rings, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     hasZ |= type.is3D;
     hasM |= type.isMeasured;
@@ -313,7 +313,7 @@ class _GeometryCollector with GeometryContent {
     Iterable<Iterable<double>> points, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     hasZ |= type.is3D;
     hasM |= type.isMeasured;
@@ -325,7 +325,7 @@ class _GeometryCollector with GeometryContent {
     Iterable<Iterable<double>> lineStrings, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     hasZ |= type.is3D;
     hasM |= type.isMeasured;
@@ -337,7 +337,7 @@ class _GeometryCollector with GeometryContent {
     Iterable<Iterable<Iterable<double>>> polygons, {
     required Coords type,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     hasZ |= type.is3D;
     hasM |= type.isMeasured;
@@ -350,7 +350,7 @@ class _GeometryCollector with GeometryContent {
     Coords? type,
     int? count,
     String? name,
-    Iterable<double>? bounds,
+    Box? bounds,
   }) {
     if (type != null) {
       hasZ |= type.is3D;